updates to mirror python files
