+++ vision/product-search/cloud-client/src/main/java/com/example/vision/ProductInProductSetManagement.java
@@ -22,7 +22,6 @@ import com.google.cloud.vision.v1p3beta1.ProductName;
 import com.google.cloud.vision.v1p3beta1.ProductSearchClient;
 import com.google.cloud.vision.v1p3beta1.ProductSet;
 import com.google.cloud.vision.v1p3beta1.ProductSetName;
-import com.google.cloud.vision.v1p3beta1.UpdateProductSetRequest;
 import com.google.protobuf.FieldMask;
 
 import java.io.IOException;
@@ -44,57 +43,9 @@ import net.sourceforge.argparse4j.inf.Subparsers;
 
 public class ProductInProductSetManagement {
 
-  // [START vision_product_search_update_product_set]
+  // [START vision_product_search_add_product_to_product_set]
   /**
-   * Update a product set.
-   *
-   * @param projectId - Id of the project.
-   * @param computeRegion - Region name.
-   * @param productSetId - Id of the product set.
-   * @param productSetDisplayName - Display name of the product set.
-   * @throws IOException - on I/O errors.
-   */
-  public static void updateProductSet(
-      String projectId, String computeRegion, String productSetId, String productSetDisplayName)
-      throws IOException {
-    ProductSearchClient client = ProductSearchClient.create();
-
-    // Get the full path of the product set.
-    String productSetPath = ProductSetName.of(projectId, computeRegion, productSetId).toString();
-
-    // Update the product set display name.
-    ProductSet productSet =
-        ProductSet.newBuilder()
-            .setName(productSetPath)
-            .setDisplayName(productSetDisplayName)
-            .build();
-
-    FieldMask updateMask = FieldMask.newBuilder().addPaths("display_name").build();
-
-    UpdateProductSetRequest request =
-        UpdateProductSetRequest.newBuilder()
-            .setProductSet(productSet)
-            .setUpdateMask(updateMask)
-            .build();
-
-    ProductSet updatedProductSet = client.updateProductSet(request);
-
-    // Display the updated set information
-    System.out.println(String.format("Product set name: %s", updatedProductSet.getName()));
-    System.out.println(
-        String.format(
-            "Product set id: %s",
-            updatedProductSet
-                .getName()
-                .substring(updatedProductSet.getName().lastIndexOf('/') + 1)));
-    System.out.println(
-        String.format("Updated product set display name: %s", updatedProductSet.getDisplayName()));
-  }
-  // [END vision_product_search_update_product_set]
-
-  // [START vision_product_search_remove_product_from_product_set]
-  /**
-   * Remove a product from a product set.
+   * Add a product to a product set.
    *
    * @param projectId - Id of the project.
    * @param computeRegion - Region name.
@@ -102,7 +53,7 @@ public class ProductInProductSetManagement {
    * @param productSetId - Id of the product set.
    * @throws IOException - on I/O errors.
    */
-  public static void removeProductFromProductSet(
+  public static void addProductToSet(
       String projectId, String computeRegion, String productId, String productSetId)
       throws IOException {
     ProductSearchClient client = ProductSearchClient.create();
@@ -113,12 +64,12 @@ public class ProductInProductSetManagement {
     // Get the full path of the product.
     String productPath = ProductName.of(projectId, computeRegion, productId).toString();
 
-    // Remove the product from the product set.
-    client.removeProductFromProductSet(productSetPath, productPath);
+    // Add the product to the product set.
+    client.addProductToProductSet(productSetPath, productPath);
 
-    System.out.println(String.format("Product removed from product set."));
+    System.out.println(String.format("Product added to product set."));
   }
-  // [END vision_product_search_remove_product_from_product_set]
+  // [END vision_product_search_add_product_to_product_set]
 
   // [START vision_product_search_list_products_in_product_set]
   /**
@@ -154,36 +105,33 @@ public class ProductInProductSetManagement {
   }
   // [END vision_product_search_list_products_in_product_set]
 
-  // [START vision_product_search_list_products]
+  // [START vision_product_search_remove_product_from_product_set]
   /**
-   * List all products.
+   * Remove a product from a product set.
    *
    * @param projectId - Id of the project.
    * @param computeRegion - Region name.
+   * @param productId - Id of the product.
+   * @param productSetId - Id of the product set.
    * @throws IOException - on I/O errors.
    */
-  public static void listProducts(String projectId, String computeRegion) throws IOException {
+  public static void removeProductFromProductSet(
+      String projectId, String computeRegion, String productId, String productSetId)
+      throws IOException {
     ProductSearchClient client = ProductSearchClient.create();
 
-    // A resource that represents Google Cloud Platform location.
-    LocationName projectLocation = LocationName.of(projectId, computeRegion);
+    // Get the full path of the product set.
+    ProductSetName productSetPath = ProductSetName.of(projectId, computeRegion, productSetId);
 
-    // List all the products available in the region.
-    for (Product product : client.listProducts(projectLocation).iterateAll()) {
-      // Display the product information
-      System.out.println(String.format("\nProduct name: %s", product.getName()));
-      System.out.println(
-          String.format(
-              "Product id: %s",
-              product.getName().substring(product.getName().lastIndexOf('/') + 1)));
-      System.out.println(String.format("Product display name: %s", product.getDisplayName()));
-      System.out.println(String.format("Product category: %s", product.getProductCategory()));
-      System.out.println("Product labels:");
-      System.out.println(
-          String.format("Product labels: %s", product.getProductLabelsList().toString()));
-    }
+    // Get the full path of the product.
+    String productPath = ProductName.of(projectId, computeRegion, productId).toString();
+
+    // Remove the product from the product set.
+    client.removeProductFromProductSet(productSetPath, productPath);
+
+    System.out.println(String.format("Product removed from product set."));
   }
-  // [END vision_product_search_list_products]
+  // [END vision_product_search_remove_product_from_product_set]
 
   public static void main(String[] args) throws Exception {
     ProductInProductSetManagement productInProductSetManagement =
@@ -194,14 +142,13 @@ public class ProductInProductSetManagement {
   public static void argsHelper(String[] args, PrintStream out) throws Exception {
     ArgumentParser parser = ArgumentParsers.newFor("").build();
     Subparsers subparsers = parser.addSubparsers().dest("command");
-    subparsers.addParser("list_products");
+
+    Subparser addProductParser = subparsers.addParser("add_product_to_product_set");
+    addProductParser.addArgument("productSetId");
+    addProductParser.addArgument("productId");
 
     Subparser listProductInProductSetParser = subparsers.addParser("list_products_in_product_set");
     listProductInProductSetParser.addArgument("productSetId");
-
-    Subparser updateProductSetParser = subparsers.addParser("update_product_set");
-    updateProductSetParser.addArgument("productSetId");
-    updateProductSetParser.addArgument("productSetDisplayName");
 
     Subparser removeProductFromProductSetParser =
         subparsers.addParser("remove_product_from_product_set");
@@ -214,22 +161,16 @@ public class ProductInProductSetManagement {
     Namespace ns = null;
     try {
       ns = parser.parseArgs(args);
-      if (ns.get("command").equals("list_products")) {
-        listProducts(projectId, computeRegion);
-      }
-      if (ns.get("command").equals("update_product_set")) {
-        updateProductSet(
-            projectId,
-            computeRegion,
-            ns.getString("productSetId"),
-            ns.getString("productSetDisplayName"));
-      }
-      if (ns.get("command").equals("remove_product_from_product_set")) {
-        removeProductFromProductSet(
+      if (ns.get("command").equals("add_product_to_product_set")) {
+        addProductToSet(
             projectId, computeRegion, ns.getString("productId"), ns.getString("productSetId"));
       }
       if (ns.get("command").equals("list_products_in_product_set")) {
         listProductsInProductSet(projectId, computeRegion, ns.getString("productSetId"));
+      }
+      if (ns.get("command").equals("remove_product_from_product_set")) {
+        removeProductFromProductSet(
+            projectId, computeRegion, ns.getString("productId"), ns.getString("productSetId"));
       }
 
     } catch (ArgumentParserException e) {
